+++ R/utilities-parameters.R
@@ -67,10 +67,19 @@ exportParametersToXLS <- function(parameters, paramsXLSpath, sheet = NULL) {
 
   for (paramIdx in seq_along(parameters)) {
     param <- parameters[[paramIdx]]
-    parameterContainerPath[[paramIdx]] <- param$parentContainer$path
-    parameterName[[paramIdx]] <- param$name
-    parameterUnits[[paramIdx]] <- param$unit
-    parameterValue[[paramIdx]] <- param$value
+    value <- param$value
+    if (!is.nan(value)) {
+      parameterContainerPath[[paramIdx]] <- param$parentContainer$path
+      parameterName[[paramIdx]] <- param$name
+      parameterUnits[[paramIdx]] <- param$unit
+      parameterValue[[paramIdx]] <- param$value
+    } else {
+      # Set to NA so these entries are removed
+      parameterContainerPath[[paramIdx]] <- NA
+      parameterName[[paramIdx]] <- NA
+      parameterUnits[[paramIdx]] <- NA
+      parameterValue[[paramIdx]] <- NA
+    }
   }
 
   output <- data.frame(
@@ -78,7 +87,9 @@ exportParametersToXLS <- function(parameters, paramsXLSpath, sheet = NULL) {
     unlist(parameterName, use.names = FALSE),
     unlist(parameterValue, use.names = FALSE),
     unlist(parameterUnits, use.names = FALSE)
-  )
+  ) %>%
+    # Remove rows for which all values are NA
+    dplyr::filter(dplyr::if_any(dplyr::everything(), ~ !is.na(.)))
 
   if (length(output) > 0) {
     colnames(output) <- c("Container Path", "Parameter Name", "Value", "Units")